+++ .gitignore
@@ -1,1 +1,4 @@
 node_modules
+/coverage
+/*.js
+/*.d.ts
