+++ package.json
@@ -4,6 +4,7 @@
   "description": "A remark markdown dialect for astronomical bulletins",
   "main": "index.js",
   "scripts": {
+    "prepare": "tsc",
     "test": "cross-env NODE_OPTIONS=--experimental-vm-modules jest"
   },
   "repository": {
@@ -25,16 +26,27 @@
     "unist-util-visit": "^4.1.1"
   },
   "devDependencies": {
+    "@tsconfig/node14": "^1.0.3",
     "@types/jest": "^29.2.5",
+    "@types/mdast": "^3.0.10",
+    "@types/unist": "^2.0.6",
     "cross-env": "^7.0.3",
     "jest": "^29.3.1",
     "mdast-util-from-markdown": "^1.2.0",
-    "prettier": "^2.8.2"
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.8.2",
+    "ts-jest": "^29.0.4",
+    "typescript": "^4.9.4"
   },
   "type": "module",
   "sideEffects": false,
   "prettier": {
     "semi": false,
     "singleQuote": true
+  },
+  "jest": {
+    "preset": "ts-jest",
+    "collectCoverage": true,
+    "coverageReporters": ["text", "cobertura"]
   }
 }
